LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-align/text-align-last-empty-inline-expected.txt

The two black line segments above should align.


FAIL content that cannot be justified should be centered when text-align-last is justify assert_equals: expected 399 but got 8
